feat: detach Controller from main.go (#14)
